Able to create new player
The new player modal is working according to the new requirements. I
still have to Set it to Add and add the pretty boarders but the error
messages are there and working. Also have started to implement a feature
that certain info will load depending on what tab you are currently in.
Example where this can be used is when Loading combo boxes for teams to
vs eachother to only Load values or new values when the tab is selected